add Scanner to get input values

+++ src/com/gmail/ncs/sapi/SimpleCalculator.java
@@ -2,15 +2,15 @@ package com.gmail.ncs.sapi;
 
 public class SimpleCalculator {
 
-    int number1;
-    int number2;
+    double number1;
+    double number2;
 
     // this is a welcome message to practice how to create / use static method
     static void welcomeMessage(){
         System.out.println("Hello this is a simple calculator. Have FUN!");
     }
 
-    public SimpleCalculator(int a, int b){
+    public SimpleCalculator(double a, double b){
         number1 = a;
         number2 = b;
     }
